requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
statsmodels>=0.13.0
mlflow>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
dvc>=3.30.0
